+++ traits/tests/test_ctraits.py
@@ -8,13 +8,16 @@
 #
 # Thanks for using Enthought open source!
 
+import sys
 import unittest
 import warnings
+import weakref
 
 from traits.constants import (
     ComparisonMode, DefaultValue, TraitKind, MAXIMUM_DEFAULT_VALUE_TYPE
 )
 from traits.ctrait import CTrait
+from traits.trait_types import Any
 
 
 def getter():
@@ -188,3 +191,24 @@ class TestCTrait(unittest.TestCase):
 
         self.assertIsInstance(trait.comparison_mode, ComparisonMode)
         self.assertEqual(trait.comparison_mode, ComparisonMode.equality)
+
+    def test_unsafe_set_value(self):
+        # Regression test for enthought/traits#832. The test below causes
+        # a segfault (on at least some systems) before the fix.
+
+        def get_handler_refcount():
+            sys.getrefcount(tr.handler)
+
+        # Anything that we can create a weakref to works here.
+        weakrefable_object = {1, 2, 3}
+
+        tr = CTrait(0)
+        tr.handler = Any(weakrefable_object)
+        finalizer = weakref.finalize(weakrefable_object, get_handler_refcount)
+        del weakrefable_object
+
+        # Reassigning the handler should trigger the finaliser.
+        self.assertTrue(finalizer.alive)
+        tr.handler = None
+        self.assertFalse(finalizer.alive)
+        self.assertIsNone(tr.handler)